feat: create query url
